vault/scripts/requirements.txt
arxiv>=2.1.0
requests>=2.28.0
biopython>=1.80
